requirements.txt
aws-cdk-lib==2.131.0
constructs>=10.0.0
boto3
botocore
types-setuptools
plotly
pandas